Refactoring of processors. Change of package. Change from services to components.
