+++ src/main/java/com/victormiranda/mani/scraper/processor/AccountProcessor.java
@@ -1,4 +1,4 @@
-package com.victormiranda.mani.scraper.service.processor;
+package com.victormiranda.mani.scraper.processor;
 
 
 import com.victormiranda.mani.bean.AccountInfo;
