+++ src/main/java/com/victormiranda/mani/scraper/processor/ptsb/PTSBAccountProcessor.java
@@ -1,23 +1,18 @@
-package com.victormiranda.mani.scraper.service.processor.ptsb;
+package com.victormiranda.mani.scraper.processor.ptsb;
 
 
 import com.victormiranda.mani.bean.AccountInfo;
-import com.victormiranda.mani.bean.ptsb.PTSBCredentials;
 import com.victormiranda.mani.scraper.bean.NavigationSession;
-import com.victormiranda.mani.scraper.service.processor.AccountProcessor;
-import com.victormiranda.mani.scraper.service.processor.BaseProcessor;
-import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
+import com.victormiranda.mani.scraper.processor.AccountProcessor;
+import com.victormiranda.mani.scraper.processor.BaseProcessor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
 import java.util.HashSet;
 import java.util.Set;
 
-import static com.victormiranda.mani.scraper.type.PTSBUrl.*;
-
-@Service
+@Component
 public class PTSBAccountProcessor extends BaseProcessor implements AccountProcessor {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(PTSBAccountProcessor.class.getName());
