+++ src/main/java/com/victormiranda/mani/scraper/service/ScraperServiceImpl.java
@@ -8,8 +8,8 @@ import com.victormiranda.mani.bean.SynchronizationResult;
 import com.victormiranda.mani.scraper.exception.LoginException;
 import com.victormiranda.mani.scraper.exception.SynchronizationException;
 import com.victormiranda.mani.scraper.bean.NavigationSession;
-import com.victormiranda.mani.scraper.service.processor.AccountProcessor;
-import com.victormiranda.mani.scraper.service.processor.LoginProcessor;
+import com.victormiranda.mani.scraper.processor.AccountProcessor;
+import com.victormiranda.mani.scraper.processor.LoginProcessor;
 import com.victormiranda.mani.scraper.type.ScraperProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
